Checking Contents of Cart When Clicking Logout
Added a condition to logout button -

If user has items in cart when clicking logout, provide a prompt asking if user wants to clear the items in cart before logging out. Then use the clearcart_Click() function to set cart to null and contents to empty.

Else, if cart does not have anything and user wants to logout, proceed to logout and set currentuser to null.