Task: Recreate Login Page
accesses the backend api/test/login. Did TDD for it as well in login.test.tsx. Displays Error Handling from API as well. To show that it uses backend it filled Landon's profile page.

User can login, and it has link to register page if user wants to register as well.